+++ src/test/java/com/ing/data/cassandra/jdbc/DbaasAstraLegacyIntegrationTest.java
@@ -19,6 +19,7 @@ import com.dtsx.astra.sdk.db.AstraDBOpsClient;
 import com.dtsx.astra.sdk.db.domain.DatabaseStatusType;
 import com.dtsx.astra.sdk.utils.TestUtils;
 import io.github.cdimascio.dotenv.Dotenv;
+import org.apache.commons.io.FileUtils;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
@@ -30,6 +31,7 @@ import org.junit.jupiter.api.condition.EnabledIf;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.File;
 import java.sql.DriverManager;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -37,29 +39,29 @@ import java.sql.SQLException;
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
 
 /**
- * Test JDBC Driver against DBaaS Astra.
+ * Test JDBC Driver against DBaaS Astra (legacy connection).
  * To run this test class, define an environment variable ASTRA_DB_APPLICATION_TOKEN containing the AstraDB token,
  * but not having any token does not block the build.
  */
 @Disabled
 @TestMethodOrder(org.junit.jupiter.api.MethodOrderer.OrderAnnotation.class)
-class DbaasAstraIntegrationTest {
+class DbaasAstraLegacyIntegrationTest {
 
-    private static final Logger LOG = LoggerFactory.getLogger(DbaasAstraIntegrationTest.class);
+    private static final Logger LOG = LoggerFactory.getLogger(DbaasAstraLegacyIntegrationTest.class);
     private static final Dotenv DOTENV = Dotenv.load();
     private static final String ASTRA_DB_TOKEN_ENV_VARIABLE = "ASTRA_DB_APPLICATION_TOKEN";
     private static final String ASTRA_DB_TOKEN = DOTENV.get(ASTRA_DB_TOKEN_ENV_VARIABLE);
-    private static final String ASTRA_DB_REGION = DOTENV.get("ASTRA_DB_REGION");
     private static final String ASTRA_DB_TOKEN_PATTERN = "Astra.*";
     private static final String DATABASE_NAME = DOTENV.get("ASTRA_DB_NAME");
     private static final String KEYSPACE_NAME = DOTENV.get("ASTRA_DB_KEYSPACE_NAME");
+    private static final String SECURE_CONNECT_BUNDLE_FILENAME = "/tmp/" + DATABASE_NAME + "_scb.zip";
 
     static CassandraConnection sqlConnection = null;
 
     @BeforeAll
     static void setupAstra() throws Exception {
         if (ASTRA_DB_TOKEN != null) {
-            LOG.debug("ASTRA_DB_APPLICATION_TOKEN is provided, AstraDB tests are executed.");
+            LOG.debug("ASTRA_DB_APPLICATION_TOKEN is provided, AstraDB tests (legacy) are executed.");
             // Set token as standard environment variable to allow TestUtils methods to work.
             System.setProperty(ASTRA_DB_TOKEN_ENV_VARIABLE, ASTRA_DB_TOKEN);
 
@@ -80,15 +82,26 @@ class DbaasAstraIntegrationTest {
             LOG.debug("Database ready.");
 
             /*
-             * Building jdbcUrl and sqlConnection, simply using the token.
+             * Download cloud secure bundle to connect to the database.
+             * - Saved in /tmp
+             * - Single region = we can use default region
+             */
+            astraDbClient
+                .database(dbId)
+                .downloadDefaultSecureConnectBundle(SECURE_CONNECT_BUNDLE_FILENAME);
+            LOG.debug("Connection bundle downloaded.");
+
+            /*
+             * Building jdbcUrl and sqlConnection.
+             * Note: Astra can be accessed with only a token (username='token').
              */
             sqlConnection = (CassandraConnection) DriverManager.getConnection(
-                "jdbc:cassandra:astra://" + DATABASE_NAME + "/" + KEYSPACE_NAME +
-                    "?token=" + ASTRA_DB_TOKEN +
-                    (ASTRA_DB_REGION != null ? "&astraregion=" + ASTRA_DB_REGION : "") +
-                    "&consistency=" + "LOCAL_QUORUM&requesttimeout=10000&connecttimeout=15000");
+                "jdbc:cassandra:dbaas:///" + KEYSPACE_NAME +
+                    "?user=token&password=" + ASTRA_DB_TOKEN +
+                    "&consistency=" + "LOCAL_QUORUM&requesttimeout=10000&connecttimeout=15000" +
+                    "&secureconnectbundle=/tmp/" + DATABASE_NAME + "_scb.zip");
         } else {
-            LOG.debug("ASTRA_DB_APPLICATION_TOKEN is not defined, skipping AstraDB tests.");
+            LOG.debug("ASTRA_DB_APPLICATION_TOKEN is not defined, skipping AstraDB tests (legacy).");
         }
     }
 
@@ -216,6 +229,8 @@ class DbaasAstraIntegrationTest {
         if (sqlConnection != null) {
             sqlConnection.close();
         }
+
+        FileUtils.deleteQuietly(new File(SECURE_CONNECT_BUNDLE_FILENAME));
     }
 
 }